4.21*: Blog List Expansion, step 9 done

+++ part4/blog-list/controllers/blogs.js
@@ -34,7 +34,19 @@ blogRouter.post("/", async (request, response) => {
 })
 
 blogRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "Invalid token" })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+  } else {
+    return response
+      .status(401)
+      .json({ error: "You are not authorized to delete this blog" })
+  }
   response.status(204).end()
 })
